fix(dbt): use NUMERIC for money so rollup sums compare exactly
assert_fct_orders_rollup_matches_items went red on 39 orders the day the
first genuinely multi-item data landed. item_count and total_quantity
matched exactly; only the amounts differed, by at most 9.095e-13 absolute
and 3.442e-16 relative -- roughly one ULP of FLOAT64. SUM() over floats is
not associative, and fct_orders' rollup and the test's re-aggregation of
fct_order_items take different execution plans, so the exact comparison
was bound to fail at random.

The bug was latent since the star schema landed: until now every order
inside the 60-day order_date window carried exactly one item, and a
single-value SUM() has no accumulation and therefore no ordering effect.

Cast item amounts to NUMERIC (exact decimal, precision 38 / scale 9)
instead of FLOAT64, so sums are exact and order-independent and the test
keeps comparing exactly rather than retreating to a tolerance -- "what
tolerance, and does it need retuning as data grows" is a question that
comes back. Verified that safe_cast keeps its fault tolerance under
NUMERIC: uncastable input and precision overflow both yield NULL, and
more than 9 decimal places rounds rather than erroring, so one dirty item
still cannot blow up a batch. quantity stays INT64; it is a count.

The change is confined to int_order_items and propagates by inference --
every downstream model is a table full rebuild, so no type migration or
--full-refresh is needed. Documented in README §5.7 and §6.1 (EN + TW).

+++ ecommerce_dbt/models/intermediate/int_order_items.sql
@@ -59,11 +59,33 @@ typed as (
         json_value(item, '$.product.condition')    as product_condition,
 
         -- 數值（safe_cast，理由見檔頭）
+        --
+        -- ⭐ 金額一律 NUMERIC，【不】用 FLOAT64（2026-08 實測後改）：
+        --   FLOAT64 的 SUM() 不滿足結合律——同一組數字換個累加順序，尾數會差一個 bit。
+        --   fct_orders 的 rollup 與 fct_order_items 的重新聚合走不同執行計畫，
+        --   於是 assert_fct_orders_rollup_matches_items 的精確比對必然隨機失敗。
+        --   實測：39 筆訂單不一致，最大相對誤差 3.442e-16（≈1 ULP，FLOAT64 epsilon 2.22e-16）。
+        --
+        --   這個問題潛伏了很久才浮現：在此之前 60 天窗內的資料每張訂單【剛好一個品項】，
+        --   單值 SUM() 沒有累加就沒有順序問題。第一批多品項資料進來當天就炸了。
+        --
+        --   NUMERIC 是精確十進位（precision 38 / scale 9），加總精確且與順序無關，
+        --   所以測試可以維持【精確比對】而不必退讓成容差比對——容差要設多少、
+        --   隨資料量成長要不要調，那是個會回來找你的決定。
+        --   金額本來就不該用浮點數，這裡只是補上這條規則。
+        --
+        --   safe_cast 的容錯行為在 NUMERIC 下同樣成立（已實測）：
+        --     轉不動（'abc'）→ NULL；超出 precision（1e400）→ NULL；
+        --     小數超過 scale 9 → 四捨五入，【不】報錯。
+        --   所以「一筆髒 item 不會炸掉整批」這個保證沒有因為換型別而失效。
+        --
+        --   quantity 維持 INT64：它是計數不是金額，整數才是正確型別
+        --   （INT64 × NUMERIC → NUMERIC，下方衍生金額不受影響）。
         safe_cast(json_value(item, '$.quantity')     as int64)   as quantity,
-        safe_cast(json_value(item, '$.unit_price')   as float64) as unit_price,
-        safe_cast(json_value(item, '$.cost_price')   as float64) as cost_price,
-        safe_cast(json_value(item, '$.discount_pct') as float64) as discount_pct,
-        safe_cast(json_value(item, '$.shipping_fee') as float64) as shipping_fee
+        safe_cast(json_value(item, '$.unit_price')   as numeric) as unit_price,
+        safe_cast(json_value(item, '$.cost_price')   as numeric) as cost_price,
+        safe_cast(json_value(item, '$.discount_pct') as numeric) as discount_pct,
+        safe_cast(json_value(item, '$.shipping_fee') as numeric) as shipping_fee
 
     from exploded
